+++ R/ppd.R
@@ -63,7 +63,7 @@ ppd <- function(f,
                 sparam = 25L,
                 parallel = TRUE,
                 cores = -1,
-                optim_method = c("DP", "DPo", "DP2", "RBFGS"),
+                optim_method = c("DP", "DPo", "RBFGS"),
                 max_iter = 20L){
 
   penalty_method <- penalty_alias(rlang::arg_match(penalty_method))